requirements.txt
numpy==1.26.4
tflite-runtime==2.14.0
pillow